requirements.txt
django
djangorestframework
django-ckeditor
django-autoslug
django-otp
django-filter
pyyaml
python-dotenv
python-decouple
psycopg2-binary
drf-yasg
drf-social-oauth2
celery
redis
firebase_admin
google-cloud-storage
google-cloud-firestore
cloudinary
pytz
pandas
fast-autocomplete[levenshtein]
infobip-api-python-sdk
django-cors-headers
social-auth-app-django
django-admin-list-filter-dropdown
django-extensions
django-celery-beat
django-import-export
setuptools
firebase-admin